# api/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the Codim API server.",
        "title": "Codim API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/courses": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Get a paginated list of courses, optionally filtered by subject",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "List courses",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Limit (default: 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset (default: 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by subject",
                        "name": "subject",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "en",
                        "example": "en",
                        "description": "Filter by language",
                        "name": "language",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": true,
                        "example": true,
                        "description": "Filter by is_active",
                        "name": "is_active",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.CourseWithTranslation"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/courses/add-translation": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Add a new translation for an existing course",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "Add a translation to an existing course",
                "parameters": [
                    {
                        "description": "Translation data",
                        "name": "translation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/courses.AddCourseTranslationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/db.CourseTranslation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/courses/create": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Create a new course",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "Create a new course",
                "parameters": [
                    {
                        "description": "Course creation data",
                        "name": "course",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/courses.CreateCourseRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/db.CourseWithTranslation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/courses/delete/{uuid}": {
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Soft delete a course by UUID (sets deleted_at timestamp)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "Delete a course",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/courses/restore": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Restore a soft-deleted course by UUID (sets deleted_at to NULL)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "Restore a deleted course",
                "parameters": [
                    {
                        "description": "Course UUID",
                        "name": "course",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/courses.IDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/courses/update": {
            "put": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Update an existing course by UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "Update a course",
                "parameters": [
                    {
                        "description": "Course update data",
                        "name": "course",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/courses.UpdateCourseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.CourseWithTranslation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/courses/{uuid}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Get a single course by UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "Get a course by UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "en",
                        "description": "Language",
                        "name": "language",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.CourseFull"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "courses.AddCourseTranslationRequest": {
            "type": "object",
            "required": [
                "course_uuid",
                "description",
                "language",
                "name"
            ],
            "properties": {
                "bullets": {
                    "type": "string",
                    "example": "Aprende lo básico\nPractica ejercicios"
                },
                "course_uuid": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "example": "Aprende los fundamentos de Python"
                },
                "language": {
                    "type": "string",
                    "example": "es"
                },
                "name": {
                    "type": "string",
                    "example": "Introducción a Python"
                }
            }
        },
        "courses.CreateCourseRequest": {
            "type": "object",
            "required": [
                "description",
                "language",
                "name",
                "price",
                "subject"
            ],
            "properties": {
                "bullets": {
                    "type": "string",
                    "example": "Learn basics\nPractice exercises"
                },
                "description": {
                    "type": "string",
                    "example": "Learn Python basics"
                },
                "difficulty": {
                    "type": "integer",
                    "example": 1
                },
                "discount": {
                    "type": "integer",
                    "example": 0
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "language": {
                    "type": "string",
                    "example": "en"
                },
                "name": {
                    "type": "string",
                    "example": "Introduction to Python"
                },
                "price": {
                    "type": "integer",
                    "example": 99
                },
                "subject": {
                    "type": "string",
                    "example": "Programming"
                }
            }
        },
        "courses.IDRequest": {
            "type": "object",
            "required": [
                "uuid"
            ],
            "properties": {
                "uuid": {
                    "type": "string"
                }
            }
        },
        "courses.UpdateCourseRequest": {
            "type": "object",
            "required": [
                "language",
                "uuid"
            ],
            "properties": {
                "bullets": {
                    "type": "string",
                    "example": "Learn basics\nPractice exercises"
                },
                "description": {
                    "type": "string",
                    "example": "Learn Python basics"
                },
                "difficulty": {
                    "type": "integer",
                    "example": 1
                },
                "discount": {
                    "type": "integer",
                    "example": 0
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "language": {
                    "type": "string",
                    "example": "en"
                },
                "name": {
                    "type": "string",
                    "example": "Introduction to Python"
                },
                "price": {
                    "type": "integer",
                    "example": 99
                },
                "subject": {
                    "type": "string",
                    "example": "Programming"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "db.CourseFull": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "difficulty": {
                    "type": "integer"
                },
                "discount": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "lessons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.LessonFull"
                    }
                },
                "modified_at": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "subject": {
                    "type": "string"
                },
                "translation": {
                    "$ref": "#/definitions/db.CourseTranslation"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "db.CourseTranslation": {
            "type": "object",
            "properties": {
                "bullets": {
                    "type": "string"
                },
                "course_uuid": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "db.CourseWithTranslation": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "difficulty": {
                    "type": "integer"
                },
                "discount": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "modified_at": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "subject": {
                    "type": "string"
                },
                "translation": {
                    "$ref": "#/definitions/db.CourseTranslation"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "db.ExerciseTranslation": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "exercise_uuid": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "db.ExerciseType": {
            "type": "string",
            "enum": [
                "quiz",
                "code"
            ],
            "x-enum-varnames": [
                "ExerciseTypeQuiz",
                "ExerciseTypeCode"
            ]
        },
        "db.ExerciseWithTranslation": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "data": {
                    "type": "object"
                },
                "deleted_at": {
                    "type": "string"
                },
                "lesson_uuid": {
                    "type": "string"
                },
                "modified_at": {
                    "type": "string"
                },
                "order_index": {
                    "type": "integer"
                },
                "reward": {
                    "type": "integer"
                },
                "translation": {
                    "$ref": "#/definitions/db.ExerciseTranslation"
                },
                "type": {
                    "$ref": "#/definitions/db.ExerciseType"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "db.LessonFull": {
            "type": "object",
            "properties": {
                "course_uuid": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "exercises": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.ExerciseWithTranslation"
                    }
                },
                "is_public": {
                    "type": "boolean"
                },
                "modified_at": {
                    "type": "string"
                },
                "order_index": {
                    "type": "integer"
                },
                "translation": {
                    "$ref": "#/definitions/db.LessonTranslation"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "db.LessonTranslation": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "lesson_uuid": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "errors.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "error": {
                    "type": "string",
                    "example": "Error message describing what went wrong"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "Authentication is performed via HTTP-only cookie named \"auth_token\" containing a JWT token. The cookie is automatically set on successful signup/login and cleared on logout.",
            "type": "apiKey",
            "name": "auth_token",
            "in": "cookie"
        }
    }
}